user-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
sqlalchemy>=2.0.22
pydantic>=2.4.2
pydantic-settings>=2.0.3
alembic>=1.12.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0
bcrypt>=4.0.1
PyJWT>=2.8.0